Refactor node API calls.
